+++ pages/telasIniciais/esqueciSenha/enviarCod/enviarCod.jsx
@@ -1,11 +1,11 @@
 import { StatusBar } from 'expo-status-bar';
 import { Text, View, TextInput} from 'react-native';
 
-import globalStyles from '../../../styles/globalStyles.js'
-import recorveryStyles from '../../../styles/recorveryStyles.js';
+import globalStyles from '../../../../styles/globalStyles.js'
+import recorveryStyles from '../../../../styles/recorveryStyles.js';
 
-import Botao from '../../../components/botao/botao';
-import LogoFacilita from '../../../components/logoFacilita/logoFacilita.jsx';
+import Botao from '../../../../components/botao/botao';
+import LogoFacilita from '../../../../components/logoFacilita/logoFacilita.jsx';
 
 
 function EnviarCod({navigation}){
