+++ pages/telasIniciais/esqueciSenha/enviarEmail/enviarEmail.jsx
@@ -1,12 +1,12 @@
 import { StatusBar } from 'expo-status-bar';
 import { Text, View, TextInput} from 'react-native';
 
-import globalStyles from '../../../styles/globalStyles.js'
-import recorveryStyles from '../../../styles/recorveryStyles.js';
+import globalStyles from '../../../../styles/globalStyles.js'
+import recorveryStyles from '../../../../styles/recorveryStyles.js';
 
-import Botao from '../../../components/botao/botao';
-import LogoFacilita from '../../../components/logoFacilita/logoFacilita.jsx';
-import ReturnArrow from '../../../components/returnArrow/returnArrow.jsx';
+import Botao from '../../../../components/botao/botao';
+import LogoFacilita from '../../../../components/logoFacilita/logoFacilita.jsx';
+import ReturnArrow from '../../../../components/returnArrow/returnArrow.jsx';
 
 function EnviarEmail({navigation}){
     return(
