+++ pages/telasIniciais/esqueciSenha/novaSenha/novaSenha.jsx
@@ -2,11 +2,11 @@ import { StatusBar } from 'expo-status-bar';
 import { Text, View, TextInput} from 'react-native';
 import { useState } from 'react';
 
-import globalStyles from '../../../styles/globalStyles.js'
-import recorveryStyles from '../../../styles/recorveryStyles.js';
+import globalStyles from '../../../../styles/globalStyles.js'
+import recorveryStyles from '../../../../styles/recorveryStyles.js';
 
-import Botao from '../../../components/botao/botao';
-import LogoFacilita from '../../../components/logoFacilita/logoFacilita.jsx';
+import Botao from '../../../../components/botao/botao';
+import LogoFacilita from '../../../../components/logoFacilita/logoFacilita.jsx';
 
 
 function NovaSenha({navigation}){
